+++ packages/splitio-adapter/src/adapter.ts
